# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseChannel": {
        "description": "Contract ``case.channel`` subset (``whatsapp``/``video`` are out of scope).",
        "enum": [
          "app_chat",
          "web_chat",
          "phone",
          "email"
        ],
        "title": "CaseChannel",
        "type": "string"
      },
...
      "CaseStatus": {
        "description": "Stored state machine of a case (see ``Case``).",
        "enum": [
          "routing",
          "queued",
          "assigned",
          "in_progress",
          "in_call",
          "to_call",
          "awaiting_approval",
          "closed"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
...
      "CreateCustomerSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "anyOf": [
              {
                "enum": [
                  "app_chat",
                  "web_chat"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Default app_chat; ignored while the customer has an open case.",
            "title": "Channel"
          },
          "customerId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Customerid",
            "type": "string"
          }
        },
        "required": [
          "customerId"
        ],
        "title": "CreateCustomerSessionRequest",
        "type": "object"
      },
      "CustomerConversation": {
        "properties": {
          "agentName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Assignee first name while with_agent.",
            "title": "Agentname"
          },
          "caseId": {
            "title": "Caseid",
            "type": "string"
          },
          "channel": {
            "$ref": "#/components/schemas/CaseChannel"
          },
          "closedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closedat"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "lastSequence": {
            "description": "Highest sequence among customer-visible turns.",
            "title": "Lastsequence",
            "type": "integer"
          },
          "openedAt": {
            "format": "date-time",
            "title": "Openedat",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CustomerConversationStatus"
          }
        },
        "required": [
          "caseId",
          "status",
          "channel",
          "language",
          "openedAt",
          "closedAt",
          "agentName",
          "lastSequence"
        ],
        "title": "CustomerConversation",
        "type": "object"
      },
      "CustomerConversationResponse": {
        "properties": {
          "conversation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerConversation"
              },
              {
                "type": "null"
              }
            ]
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/CustomerTurn"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "conversation",
          "turns"
        ],
        "title": "CustomerConversationResponse",
        "type": "object"
      },
      "CustomerConversationStatus": {
        "description": "Customer-facing projection of ``CaseStatus``.",
        "enum": [
          "waiting_agent",
          "with_agent",
          "closed"
        ],
        "title": "CustomerConversationStatus",
        "type": "string"
      },
      "CustomerLocale": {
        "enum": [
          "es-CO",
          "es-MX",
          "es-AR",
          "pt-BR"
        ],
        "title": "CustomerLocale",
        "type": "string"
      },
...
      "CustomerSelf": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "locale": {
            "$ref": "#/components/schemas/CustomerLocale"
          }
        },
        "required": [
          "id",
          "displayName",
          "locale",
          "language"
        ],
        "title": "CustomerSelf",
        "type": "object"
      },
      "CustomerSessionResponse": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/CaseChannel"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSelf"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expiresAt",
          "customer",
          "channel"
        ],
        "title": "CustomerSessionResponse",
        "type": "object"
      },
      "CustomerTurn": {
        "properties": {
          "authorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorname"
          },
          "authorRole": {
            "$ref": "#/components/schemas/CustomerTurnAuthor"
          },
          "clientMessageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only on the customer's own messages.",
            "title": "Clientmessageid"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "message",
              "notice"
            ],
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "sequence": {
            "description": "Case sequence (customer-visible turns may skip numbers).",
            "title": "Sequence",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sequence",
          "kind",
          "authorRole",
          "authorName",
          "text",
          "language",
          "createdAt",
          "clientMessageId"
        ],
        "title": "CustomerTurn",
        "type": "object"
      },
      "CustomerTurnAuthor": {
        "description": "Customer-facing author: automated tiers are shown as one \"bot\".",
        "enum": [
          "customer",
          "analyst",
          "bot",
          "system"
        ],
        "title": "CustomerTurnAuthor",
        "type": "string"
      },
...
      "Language": {
        "enum": [
          "es",
          "pt"
        ],
        "title": "Language",
        "type": "string"
      },
...
      "PostCustomerTurnRequest": {
        "additionalProperties": false,
        "properties": {
          "clientMessageId": {
            "description": "Client-generated id (UUID v4); also the Idempotency-Key header.",
            "examples": [
              "0b8f6a52-6f0e-4c1e-9d55-2f5a0c7d9e10"
            ],
            "maxLength": 64,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9-]+$",
            "title": "Clientmessageid",
            "type": "string"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "clientMessageId"
        ],
        "title": "PostCustomerTurnRequest",
        "type": "object"
      },
      "PostCustomerTurnResponse": {
        "properties": {
          "caseCreated": {
            "title": "Casecreated",
            "type": "boolean"
          },
          "conversation": {
            "$ref": "#/components/schemas/CustomerConversation"
          },
          "turn": {
            "$ref": "#/components/schemas/CustomerTurn"
          }
        },
        "required": [
          "turn",
          "conversation",
          "caseCreated"
        ],
        "title": "PostCustomerTurnResponse",
        "type": "object"
      },
...
      "ProblemCode": {
        "enum": [
          "invalid_credentials",
          "mfa_invalid",
          "mfa_challenge_invalid",
          "account_locked",
          "unauthenticated",
          "session_expired",
          "forbidden",
          "not_found",
          "method_not_allowed",
          "conflict",
          "concurrent_update",
          "invalid_transition",
          "case_not_assigned",
          "case_closed",
          "channel_not_supported",
          "idempotency_conflict",
          "invalid_value",
          "policy_violation",
          "validation_error",
          "domain_error",
          "application_error",
          "invalid_topic",
          "invalid_message",
          "http_error",
          "internal_error"
        ],
        "title": "ProblemCode",
        "type": "string"
      },
      "ProblemDetails": {
        "additionalProperties": true,
        "description": "RFC 7807 problem. ``code`` is the stable machine identifier clients branch on.\n\nThe optional members below are the documented extensions; a domain error may add other\nstructured details (e.g. ``channel`` on ``channel_not_supported``), hence\n``additionalProperties``.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "currentStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseStatus"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "invalid_transition, case_closed: the case status now."
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "validation_error: one entry per invalid field.",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "remainingAttempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "invalid_credentials, mfa_invalid: failed attempts left before the lock.",
            "title": "Remainingattempts"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Requestid"
          },
          "requiredRoles": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StaffRole"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "forbidden: roles that may perform the action.",
            "title": "Requiredroles"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "examples": [
              "urn:cc-platform:problem:account_locked"
            ],
            "title": "Type",
            "type": "string"
          },
          "unlockAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "account_locked: when the lock ends.",
            "title": "Unlockat"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "StaffRole": {
        "description": "Platform roles. They combine: one person may hold several (brief §1).",
        "enum": [
          "analyst",
          "supervisor",
          "automation",
          "admin"
        ],
        "title": "StaffRole",
        "type": "string"
      },
...
      "ValidationIssue": {
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      }
    },
    "securitySchemes": {
      "CustomerToken": {
        "description": "Customer token returned by POST /api/v1/customer/sessions (simulator).",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionToken": {
        "description": "Session token returned by POST /api/v1/auth/mfa.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API of the LATAM Bank contact-center platform (transaction-dispute intake). Errors are RFC 7807 `application/problem+json` with a stable `code`. Realtime: WebSocket `/api/v1/ws?token=…` with `{type,id,occurredAt,data}` envelopes.",
    "title": "CC Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/customer/conversation": {
      "get": {
        "operationId": "customer_get_conversation",
        "parameters": [
          {
            "in": "query",
            "name": "afterSequence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aftersequence"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerConversationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "security": [
          {
            "CustomerToken": []
          }
        ],
        "summary": "My current conversation (open case, else the last closed one) and its turns",
        "tags": [
          "customer"
        ]
      }
    },
    "/api/v1/customer/conversation/turns": {
      "post": {
        "description": "Idempotent on `clientMessageId` (= `Idempotency-Key`): a retry with the same text answers 200 with `Idempotent-Replayed: true`.",
        "operationId": "customer_post_turn",
        "parameters": [
          {
            "description": "Must equal the body `clientMessageId`; a retry with it is a replay.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Must equal the body `clientMessageId`; a retry with it is a replay.",
              "maxLength": 64,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCustomerTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCustomerTurnResponse"
                }
              }
            },
            "description": "Replay of an already-sent message"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCustomerTurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "security": [
          {
            "CustomerToken": []
          }
        ],
        "summary": "Write to the bank: opens a case when none is open, else appends to it",
        "tags": [
          "customer"
        ]
      }
    },
...
    "/api/v1/customer/sessions": {
      "post": {
        "operationId": "customer_create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Problem details (RFC 7807)"
          }
        },
        "summary": "Start a customer session (channel identity = app/web session)",
        "tags": [
          "customer"
        ]
      }
    },
...
  }
}